+++ ebpub/ebpub/accounts/views.py
@@ -35,9 +35,12 @@ import datetime
 ###########################
 
 def login(request, custom_message=None, force_form=False, initial_email=None):
-    # custom_message is a string to display at the top of the login form.
-    # force_form is used when you want to force display of the original
-    # form (regardless of whether it's a POST request).
+    """
+    View for logging in.
+    custom_message is a string to display at the top of the login form.
+    force_form is used when you want to force display of the original
+    form (regardless of whether it's a POST request).
+    """
 
     # If the user is already logged in, redirect to the dashboard.
     if not request.user.is_anonymous():
@@ -136,11 +139,7 @@ class BadHash(Exception):
         self.response = response
 
 def send_confirmation_and_redirect(request, email, task):
-    if settings.DEBUG:
-        url = utils.verification_url(email, task)
-        return http.HttpResponse('<a href="%s">Click here to simulate the e-mail confirmation</a>.' % url)
-    else:
-        utils.send_verification_email(email, task)
+    utils.send_verification_email(email, task)
     return http.HttpResponseRedirect('/accounts/email-sent/')
 
 def confirm_request_hash(request, task):